Add modular evaluator
