+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod uci;
 pub mod board;
 pub mod movegen;
 pub mod search;
+pub mod eval;
 pub mod zobrist;
 pub mod tt;
 pub mod params;
